requirements.txt
requests
pyTelegramBotAPI
psutil
python-dotenv
